+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Workbench as requirement for Clutter pieces missing a workstation requirement.
+
 ## [0.3.0] - 2024-05-04
 
 ### Changed